added generating page when portal is in the middle of environment generation, continue refactor, make posts more secure

+++ helpers/methods.rb
@@ -36,12 +36,11 @@ module Validation
   def check_error_file
     File.delete(GENERATION_FILE) if File.exists? GENERATION_FILE
     if File.exists? ERROR_FILE
-      Writer.write_response(ERROR_FILE)
-      File.delete(INFO_FILE)
+      File.delete(INFO_FILE) if File.exists? INFO_FILE
       mail = Mail.deliver do
         to EMAIL['emails']
-        from 'Portal V4 [NAME] <portalv4[name]@instructure.com>'
-        subject 'Portal Error [Name]'
+        from 'Portal V4 <portalv4@instructure.com>'
+        subject 'Portal Error'
         html_part do
           body File.read('/home/hudson/logs/sinatra_server_log.txt')
         end
@@ -51,5 +50,15 @@ module Validation
       return true
     end
   end
+  
+  def validate_patchset(input)
+    /^\d+\/\d+\/\d+$/.match(input)
+  end
+
+  def validate_gerrit_url(input)
+    /^git fetch ssh:\/\/[a-z0-9]+@[a-z0-9]+.[a-z0-9]+.com:29418\/[a-z0-9]*.*refs\/changes\/\d+\/\d+\/\d+ && git (checkout|cherry-pick) FETCH_HEAD$/.match(input)
+  end
   module_function :check_error_file
+  module_function :validate_patchset
+  module_function :validate_gerrit_url
 end
